requirements.txt
langchain
ipykernel
python-dotenv
langchain_community
pypdf
bs4
arxiv
pyMuPDF
wikipedia
langchain-text-splitters
lxml
langchain_community 
sentence-transformers
langchain_huggingface
faiss-cpu
langchain_chroma
python-dotenv
langchain
langchain_groq
langchain_core
fastapi
langserve
sse_starlette
torch
validators
youtube_transcript_api
unstructured
pytube
numexpr
langchain_nvidia_ai_endpoints